+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- The directory `target/` inside local crates won't be copied into the build
+  anymore.
+
 ## [0.3.2] - 2019-10-08
 
 ### Fixed