+++ CHANGELOG.md
@@ -10,7 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The directory `target/` inside local crates won't be copied into the build
   anymore.
 - Symbolic links will be followed instead of copied as links.
-  Better error messages will be given for broken symbolic links.
+
+### Fixed
+
+- Copying broken symbolic links will now include the path of the link.
 
 ## [0.3.2] - 2019-10-08